User Guide

User Guide

This user guide describes the basic structures and functions to define an optimisation problem, to solve the problem and to analyse the result. If you want to use JuMP to describe the problem, see the JuMP Interface section.

Model

The problem data, user settings and workspace variables are all stored in a Model. To get started define an empty model:

m = COSMO.Model()

To initialize the model with an optimisation problem we need to define three more things:

the objective function, i.e. the matrix P and the vector q in $\frac{1}{2}x^\top P x + q^\top x$

an array of constraints

a Settings object that specifies how COSMO solves the problem (optional)

Objective Function

To set the objective function of your optimisation problem simply define the square positive semidefinite matrix $P \in \mathrm{R}^{n\times n}$ and the vector $q \in \mathrm{R}^{n}$. You might have to transform your optimisation problem for this step.

Constraints

COSMO.Constraint — Type.

Constraint(A, b, convex_set, dim = 0, indices = 0:0)

Creates a COSMO constraint: Ax + b ∈ convex_set.

By default the following convex sets are supported: ZeroSet, Nonnegatives, SecondOrderCone, PsdCone, PsdConeTriangle.

Examples

julia> Constraint([1 0;0 1], zeros(2), COSMO.Nonnegatives)
Constraint
Size of A: (2, 2)
ConvexSet: COSMO.Nonnegatives

The optional arguments dim and indices can be used to specify A and b for subparts of variable x. If x has dimension dim = 4, then x[2] and x[3] can be constrained to the zero cone in the following way:

Examples

julia> c = Constraint([1 0;0 1], zeros(2), COSMO.ZeroSet, 4, 2:3)
Constraint
Size of A: (2, 4)
ConvexSet: COSMO.ZeroSet

Notice that extra columns of A have been added automatically.

julia>Matrix(c.A)
2×4 Array{Float64,2}:
0.0  1.0  0.0  0.0
0.0  0.0  1.0  0.0

Convex Sets

COSMO.ZeroSet — Type.

ZeroSet(dim)

Creates the zero set $\{ 0 \}^{dim}$ of dimension dim. If x ∈ ZeroSet then all entries of x are zero.

COSMO.Nonnegatives — Type.

Nonnegatives(dim)

Creates the nonnegative orthant $\{ x \in \mathbb{R}^{dim} : x \ge 0 \}$ of dimension dim.

COSMO.SecondOrderCone — Type.

SecondOrderCone(dim)

Creates the second-order cone (or Lorenz cone) $\{ (t,x) \in \mathrm{R}^{dim} : || x ||_2 \leq t \}$.

COSMO.PsdCone — Type.

PsdCone(dim)

Creates the cone of symmetric positive semidefinite matrices $\mathcal{S}_+^{dim}`. The entries of the matrix `X` are stored column-by-column in the vector `x` of dimension `dim`. Accordingly$X \in \mathbb{S}+ \Rightarrow x \in \mathcal{S}+^{dim}$, where$X = \text{mat}(x)``.

COSMO.PsdConeTriangle — Type.

PsdConeTriangle(dim)

Creates the cone of symmetric positive semidefinite matrices. The entries of the upper-triangular part of matrix X are stored in the vector x of dimension dim. A $r \times r$ matrix has $r(r+1)/2$ upper triangular elements and results in a vector of $\mathrm{dim} = r(r+1)/2$.

Examples

The matrix

\[\begin{bmatrix} x_1 & x_2 & x_4\\ x_2 & x_3 & x_5\\ x_4 & x_5 & x_6 \end{bmatrix}\]

is transformed to the vector $[x_1, x_2, x_3, x_4, x_5, x_6]^\top$ with corresponding constraint PsdConeTriangle(6).

Settings

Settings can be specified using the COSMO.Settings struct. The following settings are available:

Argument	Description	Values (default)

rho	ADMM rho step	0.1

sigma	ADMM sigma step	1e-6

alpha	Relaxation parameter	1.6

eps_abs	Absolute residual tolerance	1e-4

eps_rel	Relative residual tolerance	1e-4

eps_prim_inf	Primal infeasibility tolerance	1e-4

eps_dual_inf	Dual infeasibility tolerance	1e-4

max_iter	Maximum number of iterations	2500

verbose	Verbose printing	false

verbose_timing	Verbose timing	false

check_termination	Check termination interval	40

check_infeasibility	Check infeasibility interval	40

scaling	Number of scaling iterations	10

adaptive_rho	Automatic adaptation of step size parameter	true

time_limit	set solver time limit in s	0.0

For more low-level settings, see the Settings type definition in /src/settings.jl.

Results

After attempting to solve the problem, COSMO will return a result object with the following fields:

Fieldname	Type	Description

x	Vector{Float64}	Primal variable

y	Vector{Float64}	Dual variable

s	Vector{Float64}	(Primal) set variable

obj_val	Float64	Objective value

iter	Int64	Number of iterations

status	Symbol	Solution status

info	COSMO.ResultInfo	Struct with more information

times	COSMO.ResultTimes	Struct with several measured times

Status Codes

COSMO will return one of the following statuses:

Status Code	Description

:Solved	A optimal solution was found

:Unsolved	Default value

:Max_iter_reached	Solver reached iteration limit (set with Settings.max_iter)

:Time_limit_reached	Solver reached time limit (set with Settings.time_limit)

:Primal_infeasible	Problem is primal infeasible

:Dual_infeasible	Problem is dual infeasible

Timings

If settings.verbose_timing is set to true, COSMO will report the following times in result.times:

Time Name	Description

solver_time	Total time used to solve the problem

setup_time	Setup time = graph_time + factor_time

graph_time	Time used to perform chordal decomposition

factor_time	Time used to factor the system of linear equations

iter_time	Time spent in iteration loop

proj_time	Time spent in projection functions

post_time	Time used for post processing

It holds: solver_time = setup_time+ iter_time + post_time, setup_time = graph_time+ factor_time, proj_time is a subset of iter_time.