API Reference

API Reference

Model

COSMO.assemble! — Function.

assemble!(model, P, q, constraint(s), [settings, x0, y0, s0])

Assembles a COSMO.Model with a cost function defind by P and q, and a number of constraints.

The positive semidefinite matrix P and vector q are used to specify the cost function of the optimization problem:

min   1/2 x'Px + q'x
s.t.  Ax + b ∈ C

constraints is a COSMO.Constraint or an array of COSMO.Constraint objects that are used to describe the constraints on x.

The optinal arguments x0, s0, and y0 can be used to provide the solver with warm starting values for the primal variable x, the primal slack variable s and the dual variable y. The optinal argument settings can be used to pass custom solver settings.

COSMO.set! — Function.

set!(model, P, q, A, b, convex_sets, [settings])

Sets model data directly based on provided fields.

COSMO.warm_start_primal! — Function.

warm_start_primal!(model, x0, [ind])

Provides the COSMO.Model with warm starting values for the primal variable x. ind can be used to warm start certain components of x.

COSMO.warm_start_slack! — Function.

warm_start_slack!(model, s0, [ind])

Provides the COSMO.Model with warm starting values for the primal slack variable s. ind can be used to warm start certain components of s.

COSMO.warm_start_dual! — Function.

warm_start_dual!(model, y0, [ind])

Provides the COSMO.Model with warm starting values for the dual variable y. ind can be used to warm start certain components of y.

Types

COSMO.Settings — Type.

Settings(;arg=val)

Creates a COSMO settings object that is used to pass user settings to the solver.

Argument	Description	Values (default)

rho	ADMM rho step	0.1

sigma	ADMM sigma step	1e-6

alpha	Relaxation parameter	1.6

eps_abs	Absolute residual tolerance	1e-4

eps_rel	Relative residual tolerance	1e-4

eps_prim_inf	Primal infeasibility tolerance	1e-4

eps_dual_inf	Dual infeasibility tolerance	1e-4

max_iter	Maximum number of iterations	2500

verbose	Verbose printing	false

verbose_timing	Verbose timing	false

check_termination	Check termination interval	40

check_infeasibility	Check infeasibility interval	40

scaling	Number of scaling iterations	10

adaptive_rho	Automatic adaptation of step size parameter	true

time_limit	set solver time limit in s	0

COSMO.Result — Type.

Result{T <: AbstractFloat}

Object returned by the COSMO solver after calling optimize!(model). It has the following fields:

Fieldname	Type	Description

x	Vector{T}	Primal variable

y	Vector{T}	Dual variable

s	Vector{T}	(Primal) set variable

obj_val	T	Objective value

iter	Int64	Number of iterations

status	Symbol	Solution status

info	COSMO.ResultInfo	Struct with more information

times	COSMO.ResultTimes	Struct with several measured times

COSMO.Model — Type.

Model()

Initializes an empty COSMO model that can be filled with problem data using assemble!(model, P, q,constraints, [settings]).